requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2022.6.15
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-extra-fields==3.0.2
django-filter==21.1
django-smtp-ssl==1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-extra-fields==3.4.0
gunicorn==20.1.0
Pillow==2.2.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
whitenoise==6.0.0
